Update README.md
Hi,

This is my personal project page. It is deployed on the web via Github Pages and can be visited using this link: https://joeymonachino.github.io/smitesociety/

This project page was built from scratch and contains Javascript (ReactJS), HTML, CSS and a handful of libraries. There is some drag and drop responsivity and interactivity within the Tierlist page as well as javascript logic to display the Randomizer section.

This is all a work in progress and I plan to keep moving forward with more ideas, but this is a foundation for me to build upon.